requirements.txt

flask
langchain
langchain_groq
langchain_community
pypdf
python-dotenv
